# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Course Selection Shared API
  version: 1.0.0
  description: Shared HTTP contract for recommendation, enrollment, waitlist recomputation, and trace modules.
servers:
  - url: http://localhost:8000
paths:
...
  /api/enroll:
    post:
      operationId: enrollCourse
      tags: [enrollment]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
      responses:
        '200':
          description: Deterministic enrollment decision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentDecision'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/StateConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/student/status:
    get:
      operationId: getStudentStatus
      tags: [enrollment]
      parameters:
        - in: query
          name: student_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current enrollment and waitlist state for one student.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentStatusResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Student, course, trace, or scenario was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    StateConflictError:
      description: Current state does not allow the requested operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    RuleCheckResult:
      type: object
      additionalProperties: false
      required: [rule, passed, reason, related_course_id]
      properties:
        rule:
          type: string
          enum: [DUPLICATE, PREREQUISITE, TIME_CONFLICT]
        passed: {type: boolean}
        reason: {type: string, minLength: 1}
        related_course_id:
          type: [string, 'null']
    EnrollmentRequest:
      type: object
      additionalProperties: false
      required: [student_id, course_id]
      properties:
        student_id: {type: string, minLength: 1}
        course_id: {type: string, minLength: 1}
        recommendation_trace_id:
          type: [string, 'null']
    EnrollmentDecision:
      type: object
      additionalProperties: false
      required: [trace_id, student_id, course_id, rule_decision, capacity_available, status, waitlist_position, checks]
      properties:
        trace_id: {type: string, minLength: 1}
        student_id: {type: string, minLength: 1}
        course_id: {type: string, minLength: 1}
        rule_decision:
          type: string
          enum: [PASS, BLOCK]
        capacity_available: {type: boolean}
        status:
          type: string
          enum: [ENROLLED, WAITLISTED, REJECTED]
        waitlist_position:
          type: [integer, 'null']
          minimum: 1
        checks:
          type: array
          items:
            $ref: '#/components/schemas/RuleCheckResult'
    EnrollmentSummary:
      type: object
      additionalProperties: false
      required: [course_id, status]
      properties:
        course_id: {type: string}
        status:
          type: string
          enum: [ENROLLED, WAITLISTED, REJECTED]
    WaitlistEntry:
      type: object
      additionalProperties: false
      required: [student_id, course_id, position, applied_at, status, last_check_reason]
      properties:
        student_id: {type: string, minLength: 1}
        course_id: {type: string, minLength: 1}
        position: {type: integer, minimum: 1}
        applied_at: {type: string, format: date-time}
        status:
          type: string
          enum: [WAITING, PROMOTED, SKIPPED]
        last_check_reason:
          type: [string, 'null']
    StudentStatusResponse:
      type: object
      additionalProperties: false
      required: [student_id, enrollments, waitlist_entries]
      properties:
        student_id: {type: string}
        enrollments:
          type: array
          items:
            $ref: '#/components/schemas/EnrollmentSummary'
        waitlist_entries:
          type: array
          items:
            $ref: '#/components/schemas/WaitlistEntry'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, trace_id]
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, STUDENT_NOT_FOUND, COURSE_NOT_FOUND, MODEL_ERROR, STATE_CONFLICT, INTERNAL_ERROR]
            message: {type: string, minLength: 1}
            trace_id:
              type: [string, 'null']